strengthen image registry test coverage
Exercise persisted semantic invariants and exact-byte hashes. Split record, store, and service coverage by ownership.
